fix: 다운로드 후 모든 바텀시트가 닫히도록 변경
